polls/requirements.txt
chardet==3.0.4
coreapi==2.3.3
Django==2.1.7
djangorestframework==3.9.2
psycopg2==2.7.7
